Move functions from main to lib.rs
